pricestalker/requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
arrow==1.2.3
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
asyncselenium==0.0.2
attrs==22.2.0
Babel==2.9.1
binaryornot==0.4.4
briefcase==0.3.12
build==0.10.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
cookiecutter==2.1.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
magic-filter==1.0.9
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
outcome==1.2.0
packaging==23.0
pi==0.1.2
platformdirs==2.6.2
pluggy==1.0.0
psutil==5.9.4
pycparser==2.21
Pygments==2.14.0
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
pythonnet==3.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rich==13.3.1
selenium==3.141.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.1
text-unidecode==1.3
toga-core==0.3.0
toga-winforms==0.3.0
tomli_w==1.0.0
travertino==0.1.3
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.14
wsproto==1.2.0
xvfbwrapper==0.2.9
yarl==1.8.2
